refactor: 카드 유효기간 검증 로직 및 상수 분리
