+++ src/hooks/useCardValidityPeriod.tsx
@@ -1,113 +1,6 @@
 import { useState } from 'react';
-
-const PARSE_RULE = {
-  length: 2,
-} as const;
-
-const MONTH_RULE = {
-  min: 1,
-  max: 12,
-  length: 2,
-  currentMonth: new Date().getMonth() + 1,
-} as const;
-
-const YEAR_RULE = {
-  length: 2,
-  currentYear: new Date().getFullYear(),
-} as const;
-
-const ERROR_MESSAGE = {
-  MONTH_LENGTH: `MM은 ${MONTH_RULE.length}자리여야 합니다.`,
-  YEAR_LENGTH: `YY는 ${YEAR_RULE.length}자리여야 합니다.`,
-  MONTH_RANGE: `MM은 ${MONTH_RULE.min}~${MONTH_RULE.max} 사이의 숫자여야 합니다.`,
-  YEAR_RANGE: `YY는 현재 연도(${Number(
-    YEAR_RULE.currentYear.toString().slice(YEAR_RULE.length),
-  )})보다 크거나 같아야 합니다.`,
-  SAME_YEAR_MONTH_RANGE: `MM은 현재 월(${MONTH_RULE.currentMonth})보다 크거나 같고 최대 월(${MONTH_RULE.max})보다 작거나 같아야 합니다.`,
-} as const;
-
-const checkMonthLength = (month: string) => {
-  return month.length === MONTH_RULE.length;
-};
-
-const checkMonthRange = (month: number) => {
-  return month >= MONTH_RULE.min && month <= MONTH_RULE.max;
-};
-
-const checkYearLength = (year: string) => {
-  return year.length === YEAR_RULE.length;
-};
-
-const YEAR_RANGE_TYPE_RULE = {
-  same: 'same',
-  less: 'less',
-  greater: 'greater',
-} as const;
-
-const getYearRangeType = (year: number) => {
-  const currentYear = Number(
-    YEAR_RULE.currentYear.toString().slice(YEAR_RULE.length),
-  );
-
-  if (year === currentYear) {
-    return YEAR_RANGE_TYPE_RULE.same;
-  }
-
-  if (year < currentYear) {
-    return YEAR_RANGE_TYPE_RULE.less;
-  }
-
-  return YEAR_RANGE_TYPE_RULE.greater;
-};
-
-const checkMonthRangeInSameYear = (month: number) => {
-  return month >= MONTH_RULE.currentMonth && month <= MONTH_RULE.max;
-};
-
-const validateCardValidityPeriod = ({
-  month,
-  year,
-}: {
-  month: string;
-  year: string;
-}) => {
-  const newErrorMessage = {
-    month: '',
-    year: '',
-  };
-
-  const monthNumber = Number(month);
-  const yearNumber = Number(year);
-
-  const isValidMonthLength = checkMonthLength(month);
-  const isValidYearLength = checkYearLength(year);
-
-  if (!isValidMonthLength) {
-    newErrorMessage.month = ERROR_MESSAGE.MONTH_LENGTH;
-  } else {
-    if (!checkMonthRange(monthNumber)) {
-      newErrorMessage.month = ERROR_MESSAGE.MONTH_RANGE;
-    }
-  }
-
-  if (!isValidYearLength) {
-    newErrorMessage.year = ERROR_MESSAGE.YEAR_LENGTH;
-  } else {
-    const yearRangeType = getYearRangeType(yearNumber);
-    if (yearRangeType === YEAR_RANGE_TYPE_RULE.greater) {
-      newErrorMessage.year = '';
-    } else if (yearRangeType === YEAR_RANGE_TYPE_RULE.less) {
-      newErrorMessage.year = ERROR_MESSAGE.YEAR_RANGE;
-    } else if (
-      yearRangeType === YEAR_RANGE_TYPE_RULE.same &&
-      !checkMonthRangeInSameYear(monthNumber)
-    ) {
-      newErrorMessage.month = ERROR_MESSAGE.SAME_YEAR_MONTH_RANGE;
-    }
-  }
-
-  return newErrorMessage;
-};
+import { validateCardValidityPeriod } from '../validations/cardValidityPeriod';
+import { PARSE_RULE } from '../constants/cardValidityPeriod';
 
 function useCardValidityPeriod() {
   const [cardValidityPeriod, setCardValidityPeriod] = useState({
